I no longer wish to write commit messages that make sense

+++ src/play.rs
@@ -1,6 +1,7 @@
 use crate::combo::Combo;
 
+#[derive(Copy, Clone)]
 pub struct Play {
-    player_uid: u32,
+    player_uid: u8,
     trick: Combo,
 }
